Delete cluster logic
