--- code/iaas/cluster-logic/src/main/java/io/cattle/iaas/cluster/process/lock/ClusterLock.java
@@ -1,12 +1,0 @@
-package io.cattle.iaas.cluster.process.lock;
-
-import io.cattle.platform.core.model.Host;
-import io.cattle.platform.lock.definition.AbstractBlockingLockDefintion;
-
-public class ClusterLock extends AbstractBlockingLockDefintion {
-
-    public ClusterLock(Host clusterId) {
-        super("CLUSTER." + clusterId.getId());
-    }
-
-}
