+++ code/iaas/config-item/server/pom.xml
@@ -56,11 +56,6 @@
         </dependency>
         <dependency>
             <groupId>io.cattle</groupId>
-            <artifactId>cattle-iaas-cluster-logic</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cattle</groupId>
             <artifactId>cattle-iaas-service-discovery-api</artifactId>
             <version>${project.version}</version>
         </dependency>
