--- code/iaas/model/src/main/java/io/cattle/platform/core/constants/ClusterConstants.java
@@ -1,23 +1,0 @@
-package io.cattle.platform.core.constants;
-
-public class ClusterConstants {
-
-    public static final String KIND = "cluster";
-
-    public static final String ADD_REMOVE_HOST_PARAM = "hostId";
-
-    public static final String CLUSTER_HOST_MAP_REMOVE_PROCESS = "clusterhostmap.remove";
-
-    public static final String MANAGING_HOST = "managingHostId";
-
-    public static final String DISCOVERY_SPEC = "discoverySpec";
-
-    public static final String CLUSTER_SERVER_PORT = "port";
-
-    public static final String CERT_REFERENCE = "certificateId";
-
-    public static final String PROCESS_DEACTIVATE = "host.deactivate";
-
-    public static final String PROCESS_ACTIVATE = "host.activate";
-
-}
