+++ code/iaas/model/src/main/java/io/cattle/platform/core/dao/HostDao.java
@@ -14,4 +14,6 @@ public interface HostDao {
     List<? extends IpAddress> getHostIpAddresses(long hostId);
 
     Host getHostForIpAddress(long ipAddressId);
+
+    IpAddress getIpAddressForHost(Long hostId);
 }
