+++ code/implementation/host-only-network/pom.xml
@@ -15,11 +15,6 @@
         </dependency>
         <dependency>
             <groupId>io.cattle</groupId>
-            <artifactId>cattle-iaas-cluster-logic</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cattle</groupId>
             <artifactId>cattle-framework-jooq</artifactId>
             <version>${project.version}</version>
             <optional>true</optional>
