+++ code/packaging/app-config/src/main/resources/META-INF/cattle/api-server/spring-iaas-api-server-context.xml
@@ -18,6 +18,4 @@
     <bean class="io.cattle.platform.iaas.api.ippool.IpPoolAcquireActionHandler" />
     <bean class="io.cattle.platform.docker.machine.api.MachineConfigLinkHandler" />
 
-    <bean class="io.cattle.platform.iaas.api.cluster.ClusterAddHostActionHandler" />
-    <bean class="io.cattle.platform.iaas.api.cluster.ClusterRemoveHostActionHandler" />
 </beans>
