+++ code/packaging/app-config/src/main/resources/META-INF/cattle/iaas-api/spring-iaas-api-logic-context.xml
@@ -28,8 +28,6 @@
     <bean class="io.cattle.platform.iaas.api.filter.snapshot.SnapshotValidationFilter" />
     <bean class="io.cattle.platform.iaas.api.filter.ssl.CertificateCreateValidationFilter" />
 
-    <bean class="io.cattle.platform.iaas.api.cluster.HostClusterLinkFilter" />
-
     <bean class="io.cattle.platform.iaas.api.object.postinit.AccountFieldPostInitHandler" />
 
     <!-- Start Handlers -->
