--- code/packaging/app-config/src/main/resources/META-INF/cattle/system-services/cluster-logic-defaults.properties
@@ -1,4 +1,0 @@
-cluster.instance.name=Cluster Server
-cluster.insecuredocker.port=2375
-cluster.securedocker.port=2377
-cluster.config.items=cluster
