--- code/packaging/app-config/src/main/resources/META-INF/cattle/system-services/spring-cluster-context.xml
@@ -1,19 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:context="http://www.springframework.org/schema/context"
-    xmlns:aop="http://www.springframework.org/schema/aop"
-    xmlns:tx="http://www.springframework.org/schema/tx"
-    xmlns:util="http://www.springframework.org/schema/util"
-    xmlns:extension="http://cattle.io/schemas/spring/extension"
-    xsi:schemaLocation="http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.0.xsd
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.2.xsd
-        http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
-        http://cattle.io/schemas/spring/extension http://cattle.io/schemas/spring/extension-1.0.xsd
-        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
-
-    <bean class="io.cattle.iaas.cluster.service.impl.ClusterManagerImpl" />
-    <bean class="io.cattle.iaas.cluster.process.ClusterHostMapCreateHandler" />
-    <bean class="io.cattle.iaas.cluster.process.ClusterHostMapRemoveHandler" />
-
-</beans>
