+++ code/packaging/app-config/src/main/resources/META-INF/cattle/system-services/spring-core-model-system-services-context.xml
@@ -20,7 +20,6 @@
     <bean class="io.cattle.platform.core.dao.impl.NetworkDaoImpl" />
     <bean class="io.cattle.platform.core.dao.impl.StoragePoolDaoImpl" />
     <bean class="io.cattle.platform.core.dao.impl.VnetDaoImpl" />
-    <bean class="io.cattle.platform.core.dao.impl.ClusterHostMapDaoImpl" />
     <bean class="io.cattle.platform.core.dao.impl.LabelsDaoImpl" />
     <bean class="io.cattle.platform.core.dao.impl.HostDaoImpl" />
     <bean class="io.cattle.platform.core.dao.impl.VolumeDaoImpl" />
