+++ code/packaging/meta/pom.xml
@@ -207,11 +207,6 @@
         </dependency>
         <dependency>
             <groupId>io.cattle</groupId>
-            <artifactId>cattle-iaas-cluster-logic</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.cattle</groupId>
             <artifactId>cattle-iaas-service-discovery-server</artifactId>
             <version>${project.version}</version>
         </dependency>
